AI/Agents/101/Custom_Agent/pyproject.toml
[project]
name = "void"
version = "0.1.0"
description = "Custom Ai agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
